Switch matrixio-mic to threaded IRQ
The real-time constraints on the microphone array are quite tight (1).
A threaded IRQ handler runs at a higher priority than a high priority
work queue.  This helps prevent overflows of the matrixio's FPGA's
internal mic buffer.

It also simplifies the code somewhat as the kernel irq layer now takes
care of managing and triggering the threaded work instead of the driver
having to do that manually with the work queue.

(1) The FPGA's buffer is 512 frames for each period.  At the 15 MHz SPI bus
speed, it takes 4.37 ms to transfer 8 channels worth of data.  At a 96 kHz
sample rate, the period expires in 5.33 ms.  There is only 960 µs latency
to start the tranfer.